# dehaze-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "去雾系统后端 API 接口文档",
        "title": "Dehaze System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@dehaze.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/files": {
            "post": {
                "description": "文件上传",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件接口"
                ],
                "summary": "文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "表单文件对象",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "模型id",
                        "name": "modelId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SysFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "文件删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件接口"
                ],
                "summary": "文件删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "fileId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/files/check": {
            "get": {
                "description": "文件校验",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件接口"
                ],
                "summary": "文件校验",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件md5",
                        "name": "md5",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/files/download/{objectName}": {
            "get": {
                "description": "文件下载",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件接口"
                ],
                "summary": "文件下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象存储名称",
                        "name": "objectName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menus": {
            "get": {
                "description": "获取菜单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单接口"
                ],
                "summary": "菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字(菜单名称)",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1-\u003e显示；0-\u003e隐藏)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/vo.MenuVO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "新增菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单接口"
                ],
                "summary": "新增菜单",
                "parameters": [
                    {
                        "description": "菜单表单数据",
                        "name": "menuForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bo.MenuForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/menus/{id}": {
            "put": {
                "description": "修改菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单接口"
                ],
                "summary": "修改菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单表单数据",
                        "name": "menuForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bo.MenuForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单接口"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}/form": {
            "get": {
                "description": "获取菜单表单数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单接口"
                ],
                "summary": "菜单表单数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.MenuForm"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{menuId}": {
            "patch": {
                "description": "修改菜单显示状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单接口"
                ],
                "summary": "修改菜单显示状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "menuId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "显示状态(1:显示;0:隐藏)",
                        "name": "visible",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bo.MenuForm": {
            "type": "object",
            "properties": {
                "alwaysShow": {
                    "description": "【目录】只有一个子路由是否始终显示(1:是 0:否)",
                    "type": "integer"
                },
                "component": {
                    "description": "组件路径(vue页面完整路径，省略.vue后缀)",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "菜单ID",
                    "type": "integer"
                },
                "keepAlive": {
                    "description": "【菜单】是否开启页面缓存(1:开启;0:关闭)",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "path": {
                    "description": "路由路径",
                    "type": "string"
                },
                "perm": {
                    "description": "权限标识",
                    "type": "string"
                },
                "redirect": {
                    "description": "跳转路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序(数字越小排名越靠前)",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型(1-菜单；2-目录；3-外链；4-按钮权限)",
                    "type": "integer"
                },
                "visible": {
                    "description": "显示状态(1:显示;0:隐藏)",
                    "type": "integer"
                }
            }
        },
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.SysFile": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "md5": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objectName": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "vo.MenuVO": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子菜单",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.MenuVO"
                    }
                },
                "component": {
                    "description": "组件路径",
                    "type": "string"
                },
                "icon": {
                    "description": "ICON",
                    "type": "string"
                },
                "id": {
                    "description": "菜单ID",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "path": {
                    "description": "路由路径",
                    "type": "string"
                },
                "perm": {
                    "description": "按钮权限标识",
                    "type": "string"
                },
                "redirect": {
                    "description": "跳转路径",
                    "type": "string"
                },
                "sort": {
                    "description": "菜单排序(数字越小排名越靠前)",
                    "type": "integer"
                },
                "visible": {
                    "description": "菜单是否可见(1:显示;0:隐藏)",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}